various fixes and improvements to the documentation and build systems based on user feedback

+++ buildSrc/src/main/kotlin/osprey/io.kt
@@ -24,7 +24,9 @@ fun Path.createFolderIfNeeded(): Path {
 }
 
 fun Path.recreateFolder() {
-	deleteFolder()
+	if (exists()) {
+		deleteFolder()
+	}
 	createFolderIfNeeded()
 }
